4. osan tehtävät tehty

+++ osa4/blogilista/controllers/blogs.js
@@ -1,8 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -15,30 +16,45 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
 
   const body = request.body
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
   if (!blog.title || !blog.url) response.status(400).end()
 
   else response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
   
-    await Blog.findByIdAndRemove(request.params.id)
-    response
-    .status(204)
-    .end()
+    
+    const blog = await Blog.findById(request.params.id)
+    const userid = request.user._id.toString()
+
+    if (blog === undefined || blog === null) {
+      return response.status(400).json({ error: 'blog not found' })
+    }
+    console.log('blogin kirjoittaja', blog.user.toString())
+    console.log('poistava kayttaja', userid)
+
+    if (blog.user.toString() === userid) {
+      
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    }
 })
